Separating out the user selector to its own component
