+++ src/components/App/container.js
@@ -1,13 +1,8 @@
 import App from './component';
 import {connect} from 'react-redux';
-import {getUser, selectUser} from 'modules/Users'
 
-const mapStateToProps = (state) => ({
-    user: getUser(state)
-});
+const mapStateToProps = (state) => ({});
 
-const mapDispatchToProps = (dispatch) => ({
-    selectUser: (user) => dispatch(selectUser(user))
-});
+const mapDispatchToProps = (dispatch) => ({});
 
 export default connect(mapStateToProps, mapDispatchToProps)(App);
