+++ src/components/UserSelector/index.js
@@ -1,0 +1,3 @@
+import UserSelector from './container';
+
+export default UserSelector;
